Corrected issue with Submitter
Brought priority and project attributes into JobTicket because every job
has them regardless of what type of job it is.

Fixed the issue with Submitter putting QString objects into the command
line and reduced ambiguity with different uses of the word "project".
"project" refers to the particular render farm the job is meant to be
run on. "projectPath" is now "mayaProjectPath"